+++ R/AllClasses.R
@@ -49,6 +49,7 @@
 #' @slot singular_values A [`numeric`] vector giving the singular values.
 #' @slot rows A [`MultivariateResults-class`] object.
 #' @slot columns A [`MultivariateResults-class`] object.
+#' @slot extra A [`matrix`] of extra variables.
 #' @section Subset:
 #'  In the code snippets below, `x` is a `MultivariateAnalysis` object.
 #'  \describe{
@@ -68,7 +69,8 @@
     dimension = "integer",
     singular_values = "numeric",
     rows = "MultivariateResults",
-    columns = "MultivariateResults"
+    columns = "MultivariateResults",
+    extra = "matrix"
   ),
   contains = "VIRTUAL"
 )